correct args

+++ ios/Classes/DoozMeshManagerApi/Channels/Arguments/ConfigBeaconGetArguments.swift
@@ -6,5 +6,5 @@
 //
 
 struct ConfigBeaconGetArguments: BaseFlutterArguments {
-    let address: Int16
+    let address: Int
 }
